much better parallel version with mmap
If we mmap the whole file to get it in memory as a giant buffer, we can
use the much more efficient ParallelString::par_lines adapter to get
a parallel iterator that doesn't have to allocate for every single line,
and also doesn't have to check an io::Result for every line either.

Figuring out fold() + reduce() was the trickiest part of this - not
remembering the "fold" terminology off the top of my head meant I tried
messing with map_with and reduce for too long before figuring out that
fold was basically what I wanted all along.

Performance is faster than the single-threaded version, but only about
4-5x speedup even though I'm on 12 threads instead of 1.

% hyperfine --runs 5 -N 'target/release/onebrc measurements.txt'
Benchmark 1: target/release/onebrc measurements.txt
  Time (mean ± σ):     13.493 s ±  0.043 s    [User: 135.933 s, System: 0.467 s]
  Range (min … max):   13.460 s … 13.565 s    5 runs

Perf-Tag: Parallel String mmap
Perf-Mean: 13.493
Perf-Stdev: 0.043